+++ configs/seasonal_forecast_example_config.cfg
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 [parameters]
-client=fake
+client=cds
 dataset=seasonal-original-single-levels
 target_path=gs://ecmwf-output-test/seasonal-forecast/seasonal-forecast-{}-{}.nc
 partition_keys=
