+++ configs/yesterdays_surface_example.cfg
@@ -13,7 +13,7 @@
 # limitations under the License.
 
 [parameters]
-client=fake
+client=mars
 partition_keys=
     type
 target_path=all-{}.an
